Add support for setting `take_over` (#341)
The new optional Boolean parameter `take_over` for `filebeat::input`
defaults to `undef` but can be set to `true` to enable the "take over"
functionality when switching from the deprecated `log` input type to the
new `filestream` type. This allows Filebeat to take over files from the
`log` input without completely re-ingesting the files, which leads to
duplicated log entries. This feature is available in Filebeat 8.x.

This change adds very basic spec tests for the `filestream` input type
and the `take_over` logic, and updates the docs where appropriate.